move player_bar to player directory;
